requirements.txt
torch==2.0.0
torchvision==0.15.0
segmentation-models-pytorch==0.2.1
opencv-python
matplotlib
numpy
Pillow